requirements.txt
altgraph==0.17.4
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
auto-py-to-exe==2.42.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
beniget==0.4.1
bleach==6.1.0
blinker==1.4
bottle==0.12.25
bottle-websocket==0.2.9
Brotli==1.0.9
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
comm==0.1.4
command-not-found==0.3
common==0.1.2
cryptography==43.0.1
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
Eel==0.16.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.13.4
fonttools==4.29.1
fqdn==1.5.1
fs==2.4.12
fsspec==2024.3.1
future==0.18.3
gast==0.5.2
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.3
html5lib==1.1
httplib2==0.20.2
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
importlib-metadata==4.6.4
imutils==0.5.4
ipykernel==6.25.2
ipython==8.16.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-console==6.4.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lxml==4.8.0
lz4==3.1.3+dfsg
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==8.10.0
moviepy==1.0.3
mpmath==0.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.9.0.80
opencv-python==4.8.1.78
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pdfannots==0.4
pdfminer.six==20240706
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.13
planutils==0.10.15
platformdirs==3.11.0
ply==3.11
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pycparser==2.21
Pygments==2.16.1
PyGObject==3.42.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyJWT==2.3.0
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pyparsing==2.4.7
PyPDF2==3.0.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-json-logger==2.0.7
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
pyzmq==25.1.1
referencing==0.30.2
reportlab==4.1.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.9
systemd-python==234
terminado==0.17.1
testpath==0.5.0
texttable==1.6.4
threadpoolctl==3.2.0
tifffile==2023.9.26
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
triton==2.2.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==27.11.2
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.3.0
urllib3==2.0.7
wadllib==1.3.6
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
whichcraft==0.6.1
widgetsnbextension==2.0.0
yarg==0.1.9
zipp==1.0.0
zope.event==5.0
zope.interface==6.1
